+++ Modules/DiffusionImaging/FiberTracking/IODataStructures/FiberBundle/mitkFiberBundle.cpp
@@ -1012,7 +1012,7 @@ std::vector<long> mitk::FiberBundle::ExtractFiberIdSubset(DataNode *roi, DataSto
                 result.push_back(i);
 
             std::vector<long>::iterator it;
-            for (long i=0; i<children->Size(); ++i)
+            for (unsigned int i=0; i<children->Size(); ++i)
             {
                 std::vector<long> inRoi = ExtractFiberIdSubset(children->ElementAt(i), storage);
